Add `--refresh` behavior to the cache (#1057)
## Summary

This PR is an alternative approach to #949 which should be much safer.
As in #949, we add a `Refresh` policy to the cache. However, instead of
deleting entries from the cache the first time we read them, we now
check if the entry is sufficiently new (created after the start of the
command) if the refresh policy applies. If the entry is stale, then we
avoid reading it and continue onward, relying on the cache to
appropriately overwrite based on "new" data. (This relies on the
preceding PRs, which ensure the cache is append-only, and ensure that we
can atomically overwrite.)

Unfortunately, there are just a lot of paths through the cache, and
didn't data is handled with different policies, so I really had to go
through and consider the "right" behavior for each case. For example,
the HTTP requests can use `max-age=0, must-revalidate`. But for the
routes that are based on filesystem modification, we need to do
something slightly different.

Closes #945.

+++ crates/puffin-interpreter/src/interpreter.rs
@@ -11,7 +11,7 @@ use pep440_rs::Version;
 use pep508_rs::MarkerEnvironment;
 use platform_host::Platform;
 use platform_tags::{Tags, TagsError};
-use puffin_cache::{Cache, CacheBucket, CachedByTimestamp};
+use puffin_cache::{Cache, CacheBucket, CachedByTimestamp, Freshness};
 use puffin_fs::write_atomic_sync;
 
 use crate::python_platform::PythonPlatform;
@@ -272,6 +272,7 @@ impl InterpreterQueryResult {
     /// time as a cache key.
     pub(crate) fn query_cached(executable: &Path, cache: &Cache) -> Result<Self, Error> {
         let executable_bytes = executable.as_os_str().as_encoded_bytes();
+
         let cache_entry = cache.entry(
             CacheBucket::Interpreter,
             "",
@@ -281,25 +282,30 @@ impl InterpreterQueryResult {
         let modified = Timestamp::from_path(fs_err::canonicalize(executable)?.as_ref())?;
 
         // Read from the cache.
-        if let Ok(data) = fs::read(cache_entry.path()) {
-            match rmp_serde::from_slice::<CachedByTimestamp<Timestamp, Self>>(&data) {
-                Ok(cached) => {
-                    if cached.timestamp == modified {
-                        debug!("Using cached markers for: {}", executable.display());
-                        return Ok(cached.data);
-                    }
+        if cache
+            .freshness(&cache_entry, None)
+            .is_ok_and(Freshness::is_fresh)
+        {
+            if let Ok(data) = fs::read(cache_entry.path()) {
+                match rmp_serde::from_slice::<CachedByTimestamp<Timestamp, Self>>(&data) {
+                    Ok(cached) => {
+                        if cached.timestamp == modified {
+                            debug!("Using cached markers for: {}", executable.display());
+                            return Ok(cached.data);
+                        }
 
-                    debug!(
-                        "Ignoring stale cached markers for: {}",
-                        executable.display()
-                    );
-                }
-                Err(err) => {
-                    warn!(
-                        "Broken cache entry at {}, removing: {err}",
-                        cache_entry.path().display()
-                    );
-                    let _ = fs_err::remove_file(cache_entry.path());
+                        debug!(
+                            "Ignoring stale cached markers for: {}",
+                            executable.display()
+                        );
+                    }
+                    Err(err) => {
+                        warn!(
+                            "Broken cache entry at {}, removing: {err}",
+                            cache_entry.path().display()
+                        );
+                        let _ = fs_err::remove_file(cache_entry.path());
+                    }
                 }
             }
         }
